+++ server/subsonic/browsing.go
@@ -354,8 +354,8 @@ func (api *Router) buildArtistDirectory(ctx context.Context, artist *model.Artis
 	if artist.PlayCount > 0 {
 		dir.Played = &artist.PlayDate
 	}
-	dir.AlbumCount = artist.AlbumCount
-	dir.UserRating = artist.Rating
+	dir.AlbumCount = int32(artist.AlbumCount)
+	dir.UserRating = int32(artist.Rating)
 	if artist.Starred {
 		dir.Starred = &artist.StarredAt
 	}
@@ -392,8 +392,8 @@ func (api *Router) buildAlbumDirectory(ctx context.Context, album *model.Album) 
 	if album.PlayCount > 0 {
 		dir.Played = &album.PlayDate
 	}
-	dir.UserRating = album.Rating
-	dir.SongCount = album.SongCount
+	dir.UserRating = int32(album.Rating)
+	dir.SongCount = int32(album.SongCount)
 	dir.CoverArt = album.CoverArtID().String()
 	if album.Starred {
 		dir.Starred = &album.StarredAt
@@ -415,15 +415,15 @@ func (api *Router) buildAlbum(ctx context.Context, album *model.Album, mfs model
 	dir.Artist = album.AlbumArtist
 	dir.ArtistId = album.AlbumArtistID
 	dir.CoverArt = album.CoverArtID().String()
-	dir.SongCount = album.SongCount
-	dir.Duration = int(album.Duration)
+	dir.SongCount = int32(album.SongCount)
+	dir.Duration = int32(album.Duration)
 	dir.PlayCount = album.PlayCount
 	if album.PlayCount > 0 {
 		dir.Played = &album.PlayDate
 	}
-	dir.Year = album.MaxYear
+	dir.Year = int32(album.MaxYear)
 	dir.Genre = album.Genre
-	dir.UserRating = album.Rating
+	dir.UserRating = int32(album.Rating)
 	if !album.CreatedAt.IsZero() {
 		dir.Created = &album.CreatedAt
 	}
